% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{add_random_ids}
\alias{add_random_ids}
\title{Add randome IDs to data frame}
\usage{
add_random_ids(df, ID, drop = TRUE)
}
\arguments{
\item{df}{Data frame with which to add new random ID}

\item{ID}{The existing ID field.}

\item{drop}{Whether to drop the original ID field. If
ID field is NOT dropped, it will be renamed OID.
(default=TRUE)}
}
\description{
Add randome IDs to data frame
}
