% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.r
\name{getColleagueData_SQL}
\alias{getColleagueData_SQL}
\title{Return a data from data warehouse Colleague tables via a SQL query.}
\usage{
getColleagueData_SQL(
  sql,
  db = NA_character_,
  sep = ".",
  cfg = NA_character_,
  cfg_full_path = NA_character_,
  cfg_fn = NA_character_,
  cfg_path = NA_character_,
  reload = FALSE
)
}
\arguments{
\item{sql}{The query to run.}

\item{db}{The database to use. Defaults to the database specified in the configuration file.}

\item{sep}{The separator to use in the field names. Default is a '.' as in the original Colleague file.}

\item{cfg}{A YAML configuration file with `sql` section that includes `driver` `server`, `db`, and `schema_default`.}

\item{cfg_full_path}{Provide the full path to the YAML configuration file. Defaults to NA.}

\item{cfg_fn}{The file name for the YAML configuration file. Defaults to config.yml.}

\item{cfg_path}{The file path to the YAML configuration file. Defaults to '.'.}

\item{reload}{Force a reload of the config.}

\item{ext}{The extention to use for files.}
}
\value{
The requested data in a dataframe.
}
\description{
Return a data from data warehouse Colleague tables via a SQL query.
}
