% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{\%nlike any\%}
\alias{\%nlike any\%}
\title{The opposite of \%like any\%.}
\usage{
x \%nlike any\% pattern
}
\arguments{
\item{x}{Elements to return if it matches any one pattern.}

\item{pattern}{Patterns to match.}
}
\description{
The opposite of \%like any\%.
}
