% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.r
\name{mv_to_delim}
\alias{mv_to_delim}
\title{Convert a collection of multi-valued columns into delimiter-separated columns.}
\usage{
mv_to_delim(df, keys, assoc, cols, delim)
}
\arguments{
\item{df}{The source dataframe.}

\item{keys}{A list of all key columns, used to uniquely identify each "row".}

\item{assoc}{A dictionary (dataframe) of associations of multi-valued columns. These columns are collapsed together as a group.}

\item{cols}{A list of independent multi-valued columns. These columns are collapsed individually.}

\item{delim}{The separator to use between  values. This defaults to a comma followed by a space.}
}
\description{
That is, it takes this:
}
\details{
ID         Col1       Col2
-------    --------   -------
001        A          1
001        B          2
001        C          3

...and by invoking it with mv_to_delim(df, keys=c('ID'), assoc=dataframe('GRP1' = c('Col1','Col2'))),
converts it to this:

ID         Col1       Col2
-------    --------   -------
001        A, B, C    1, 2, 3
}
