% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.r
\name{setCfg}
\alias{setCfg}
\title{Set a new value in the cached YAML configuration file.}
\usage{
setCfg(
  section,
  variable,
  value,
  cfg = NA_character_,
  cfg_full_path = NA_character_,
  cfg_fn = NA_character_,
  cfg_path = NA_character_,
  reload = FALSE
)
}
\arguments{
\item{section}{The section for the new variable}

\item{variable}{The name of the new variable}

\item{value}{The value for the new variable}

\item{cfg}{A YAML configuration file with `sql` section that includes `driver` `server`, `db`, and `schema_history`.}

\item{cfg_full_path}{Provide the full path to the YAML configuration file. Defaults to NA.}

\item{cfg_fn}{The file name for the YAML configuration file. Defaults to config.yml.}

\item{cfg_path}{The file path to the YAML configuration file. Defaults to '.'.}

\item{reload}{Force a reload of the config.}
}
\value{
A configuration dataframe.
}
\description{
Set a new value in the cached YAML configuration file.
}
